mv: move ensemble posterior to posteriors. fixes ImportError. Add deprecation warnings
