+++ sbi/inference/posteriors/__init__.py
@@ -1,4 +1,5 @@
 from sbi.inference.posteriors.direct_posterior import DirectPosterior
+from sbi.inference.posteriors.ensemble_posterior import EnsemblePosterior
 from sbi.inference.posteriors.importance_posterior import ImportanceSamplingPosterior
 from sbi.inference.posteriors.mcmc_posterior import MCMCPosterior
 from sbi.inference.posteriors.rejection_posterior import RejectionPosterior
