+++ sbi/utils/__init__.py
@@ -71,4 +71,3 @@ from sbi.utils.user_input_checks import (
     validate_theta_and_x,
 )
 from sbi.utils.user_input_checks_utils import MultipleIndependent
-from sbi.utils.posterior_ensemble import NeuralPosteriorEnsemble
